Discard flake-utils dependency in favor of nixpkgs.lib.genAttrs
Flake Utils have a lot of quirks due to way it alters the flake when
resolving the final schema. Using `nixpkgs.lib.genAttrs` may seems
slightly less convinient at first, but it does not alter the flake
output in some way that would be hard to predict; each call to
forAllSystem will create the system attribute key at the expected place.
This also avoids fetching a dependency that may cascade to dependent
flakes.

In-details information about this refactor may be found at
[here](https://ayats.org/blog/no-flake-utils).

Finally this commit also wraps the package definition onto the
callPackage construct, which is the proper way to handle
stdenv.mkDerivation.